% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{.themeOptions}
\alias{.themeOptions}
\title{Get Theme Options}
\usage{
.themeOptions(theme)
}
\arguments{
\item{theme}{A character string representing the theme name.}
}
\value{
A list with `colors`, `font`, and `size` settings for the theme.
}
\description{
Retrieves options for a specified map theme, including color, font, and size settings.
Applies `default_size` if no size data is found for the theme.
}
