% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes_back.R
\docType{data}
\name{.themes}
\alias{.themes}
\title{Available Themes}
\format{
A character vector of theme names.
}
\usage{
.themes
}
\description{
A list of theme names for various city map styles.
}
\keyword{datasets}
