% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{list_themes}
\alias{list_themes}
\title{List Available Themes}
\usage{
list_themes()
}
\value{
A character vector of theme names.
}
\description{
Lists the names of all predefined themes available in the package for use in custom map styling.
}
