% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R, R/themes_back.R
\name{manage_themes}
\alias{manage_themes}
\title{Unregister or Register Themes}
\usage{
manage_themes(remove = F)

manage_themes(remove = F)
}
\arguments{
\item{remove}{Logical; if TRUE, unregister themes. If FALSE, register them.}
}
\value{
A message indicating that all themes have been unregistered.
}
\description{
Registers or unregisters themes from `rcityviews`, depending on the `remove` parameter.

Unregisters all themes from `rcityviews_extra` by iterating over each theme
and removing it from the `rcityviews` cache, bypassing the confirmation prompt.
}
\details{
This function removes all themes in `.themes` from the `rcityviews` cache, automatically selecting "Yes"
for each removal confirmation. It provides an efficient way to reset theme registrations.
}
