% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R, R/themes_back.R
\name{.themeOptions}
\alias{.themeOptions}
\title{Get Theme Options}
\usage{
.themeOptions(theme)

.themeOptions(theme)
}
\arguments{
\item{theme}{A character string representing the theme name. Must be one of the predefined theme names in `.themes`.}
}
\value{
A list with `colors`, `font`, and `size` settings for the theme.

A list with three elements:
\describe{
  \item{colors}{A list of colors for background, water, landuse, contours, streets, rails, buildings, text, and waterlines.}
  \item{font}{A list with font family, face, and scale for the text used in the theme.}
  \item{size}{A list with size parameters for borders and streets, detailing contour lines, water, streets, and rails width.}
}
}
\description{
Retrieves options for a specified map theme, including color, font, and size settings.
Applies `default_size` if no size data is found for the theme.

This function returns a list of options for a specified map theme, including color, font, and size settings.
Each theme provides unique color schemes and font choices tailored to different styles and cities.
}
