% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{manage_themes}
\alias{manage_themes}
\title{Unregister or Register Themes}
\usage{
manage_themes(remove = FALSE)
}
\arguments{
\item{remove}{Logical; if TRUE, unregister themes. If FALSE, register them.}
}
\description{
Registers or unregisters themes from `rcityviews`, depending on the `remove` parameter.
}
